--- front/src/Components/PharmList/PharmDetail.tsx
@@ -1,7 +1,0 @@
-export default function PharmDetail() {
-  return (
-    <div>
-      
-    </div>
-  );
-}
