+++ front/src/Pages/Home.tsx
@@ -1,10 +1,11 @@
 import React from "react";
+import PharmDetail from "../Components/Modal/PharmDetail";
 import PharmModal from "../Components/Modal/PharmModal";
 
 export default function Home() {
   return (
     <div>
-      <PharmModal />
+      <PharmDetail />
     </div>
   );
 }
